Simplified test fixture loading
