+++ tests/core/demo_units/procedures/01-create-temp-tables.sql
