+++ tests/core/procedures/basic-insert.sql
