+++ tests/core/procedures/table-exists.sql
