+++ tests/core/test_03_fill_holes.py
@@ -1,4 +1,4 @@
-from .helpers import (
+from ..helpers import (
     insert_line,
     insert_polygon,
     map_layer_id,
