+++ tests/core/test_05_map_faces.py
@@ -3,7 +3,7 @@
 from macrostrat.utils.timer import Timer
 from macrostrat.utils import get_logger
 
-from .helpers import (
+from ..helpers import (
     insert_line,
     map_layer_id,
     add_linework_type_to_layer,
